move old i18n version to api-v2 folder and use new i18n where possible
