+++ test/helpers/content.helper.js
@@ -3,7 +3,7 @@ import {each} from 'lodash';
 
 import i18n from '../../common/script/src/i18n';
 require('coffee-script');
-i18n.translations = require('../../website/src/libs/i18n.js').translations;
+i18n.translations = require('../../website/src/libs/api-v3/i18n').translations;
 
 export const STRING_ERROR_MSG = 'Error processing the string. Please see Help > Report a Bug.';
 export const STRING_DOES_NOT_EXIST_MSG = /^String '.*' not found.$/;
