+++ website/src/controllers/api-v2/auth.js
@@ -10,7 +10,7 @@ var FirebaseTokenGenerator = require('firebase-token-generator');
 var User = require('../../models/user').model;
 var EmailUnsubscription = require('../../models/emailUnsubscription').model;
 var analytics = utils.analytics;
-var i18n = require('./../../libs/i18n');
+var i18n = require('./../../libs/api-v2/i18n');
 
 var isProd = nconf.get('NODE_ENV') === 'production';
