+++ website/src/middlewares/locals.js
@@ -2,7 +2,7 @@ var nconf = require('nconf');
 var _ = require('lodash');
 var utils = require('../libs/utils');
 var shared = require('../../../common');
-var i18n = require('../libs/i18n');
+var i18n = require('../libs/api-v2/i18n');
 var buildManifest = require('../libs/buildManifest');
 var shared = require('../../../common');
 var forceRefresh = require('./forceRefresh');
