+++ website/src/routes/api-v2/coupon.js
@@ -3,7 +3,7 @@ var express = require('express');
 var router = new express.Router();
 var auth = require('../../controllers/api-v2/auth');
 var coupon = require('../../controllers/api-v2/coupon');
-var i18n = require('../../libs/i18n');
+var i18n = require('../../libs/api-v2/i18n');
 
 router.get('/api/v2/coupons', auth.authWithUrl, i18n.getUserLanguage, coupon.ensureAdmin, coupon.getCoupons);
 router.post('/api/v2/coupons/generate/:event', auth.auth, i18n.getUserLanguage, coupon.ensureAdmin, coupon.generateCoupons);
