+++ website/src/routes/api-v2/unsubscription.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = new express.Router();
-var i18n = require('../../libs/i18n');
+var i18n = require('../../libs/api-v2/i18n');
 var unsubscription = require('../../controllers/api-v2/unsubscription');
 
 router.get('/unsubscribe', i18n.getUserLanguage, unsubscription.unsubscribe);
